not completed

+++ bonus/src/strlen_char_bonus.c
@@ -6,7 +6,7 @@
 /*   By: alafdili <alafdili@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/01 12:36:08 by alafdili          #+#    #+#             */
-/*   Updated: 2024/04/14 14:16:31 by alafdili         ###   ########.fr       */
+/*   Updated: 2024/04/20 16:37:58 by alafdili         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	strlen_char(char *s)
 {
-	int_least64_t	counter;
+	int	counter;
 
 	counter = 0;
 	while (s[counter] && s[counter] != '\n')
